backend/requirements.txt
fastapi==0.116.1
langchain_core==0.3.72
langchain_openai==0.3.28
langgraph==0.6.3
matplotlib==3.10.5
pandas==2.3.1
pandasql==0.7.3
python-dotenv==1.1.1
uvicorn[standard]
openpyxl
xlrd
python-multipart
pydantic
python-dotenv
requests
